Added more guards

+++ BrothermanBill/stop.sh
@@ -7,20 +7,23 @@ echo "Process killed" #stops STDOUT from hanging and causing error
 
 # https://superuser.com/questions/513159/how-to-remove-systemd-services
 
-if systemctl is-active --quiet brothermanbill.service; then
-	sudo systemctl stop brothermanbill.service
-fi
+if systemctl list-unit-files | grep brothermanbill.service; then 
 
-if systemctl is-enabled --quiet brothermanbill.service; then
-	sudo systemctl disable brothermanbill.service
+	if systemctl is-active --quiet brothermanbill.service; then
+		sudo systemctl stop brothermanbill.service
+	fi
+
+	if systemctl is-enabled --quiet brothermanbill.service; then
+		sudo systemctl disable brothermanbill.service
+	fi
 fi
 
 if [ -f "/etc/systemd/system/brothermanbill.service" ] ; then
-    rm "/etc/systemd/system/brothermanbill.service"
+	rm "/etc/systemd/system/brothermanbill.service"
 fi
 
-if [ -f "/usr/lib/systemd/system/brothermanbill.service " ] ; then
-    rm "/usr/lib/systemd/system/brothermanbill.service "
+if [ -f "/usr/lib/systemd/system/brothermanbill.service" ] ; then
+	rm "/usr/lib/systemd/system/brothermanbill.service"
 fi
 
 sudo systemctl daemon-reload
